feat(rust): move storage in `ockam_identity`
