+++ implementations/rust/ockam/ockam_identity/src/v2/identities/storage/identities_repository_impl.rs
@@ -1,35 +1,27 @@
+use ockam_core::async_trait;
 use ockam_core::compat::boxed::Box;
 use ockam_core::compat::collections::BTreeMap;
 use ockam_core::compat::string::ToString;
 use ockam_core::compat::sync::Arc;
 use ockam_core::compat::vec::Vec;
-use ockam_core::errcode::{Kind, Origin};
 use ockam_core::Result;
-use ockam_core::{async_trait, Error};
 
-use super::super::super::identities::storage::storage::Storage;
 use super::super::super::identity::IdentityConstants;
 use super::super::super::models::{ChangeHistory, Identifier};
+use super::super::super::storage::{InMemoryStorage, Storage};
 use super::super::super::utils::now;
 use super::super::super::{Identity, IdentityError, IdentityHistoryComparison};
-use super::super::{AttributesEntry, IdentitiesVault, InMemoryStorage};
+use super::super::IdentitiesVault;
+use super::{
+    AttributesEntry, IdentitiesReader, IdentitiesRepository, IdentitiesWriter,
+    IdentityAttributesReader, IdentityAttributesWriter,
+};
 
-/// Repository for data related to identities: key changes and attributes
-#[async_trait]
-pub trait IdentitiesRepository:
-    IdentityAttributesReader + IdentityAttributesWriter + IdentitiesReader + IdentitiesWriter
-{
-    /// Restrict this repository as a reader for attributes
-    fn as_attributes_reader(&self) -> Arc<dyn IdentityAttributesReader>;
-
-    /// Restrict this repository as a writer for attributes
-    fn as_attributes_writer(&self) -> Arc<dyn IdentityAttributesWriter>;
-
-    /// Restrict this repository as a reader for identities
-    fn as_identities_reader(&self) -> Arc<dyn IdentitiesReader>;
-
-    /// Restrict this repository as a writer for identities
-    fn as_identities_writer(&self) -> Arc<dyn IdentitiesWriter>;
+/// Implementation of `IdentityAttributes` trait based on an underlying `Storage`
+#[derive(Clone)]
+pub struct IdentitiesStorage {
+    storage: Arc<dyn Storage>,
+    vault: Arc<dyn IdentitiesVault>, // TODO: Reconsider Vault dependency
 }
 
 #[async_trait]
@@ -49,69 +41,6 @@ impl IdentitiesRepository for IdentitiesStorage {
     fn as_identities_writer(&self) -> Arc<dyn IdentitiesWriter> {
         Arc::new(self.clone())
     }
-}
-
-/// Trait implementing read access to attributes
-#[async_trait]
-pub trait IdentityAttributesReader: Send + Sync + 'static {
-    /// Get the attributes associated with the given identity identifier
-    async fn get_attributes(&self, identity: &Identifier) -> Result<Option<AttributesEntry>>;
-
-    /// List all identities with their attributes
-    async fn list(&self) -> Result<Vec<(Identifier, AttributesEntry)>>;
-}
-
-/// Trait implementing write access to attributes
-#[async_trait]
-pub trait IdentityAttributesWriter: Send + Sync + 'static {
-    /// Set the attributes associated with the given identity identifier.
-    /// Previous values gets overridden.
-    async fn put_attributes(&self, identity: &Identifier, entry: AttributesEntry) -> Result<()>;
-
-    /// Store an attribute name/value pair for a given identity
-    async fn put_attribute_value(
-        &self,
-        subject: &Identifier,
-        attribute_name: Vec<u8>,
-        attribute_value: Vec<u8>,
-    ) -> Result<()>;
-
-    /// Remove all attributes for a given identity identifier
-    async fn delete(&self, identity: &Identifier) -> Result<()>;
-}
-
-/// Trait implementing write access to identities
-#[async_trait]
-pub trait IdentitiesWriter: Send + Sync + 'static {
-    /// Store changes if there are new key changes associated to that identity
-    /// Return an error if the current change history conflicts with the persisted one
-    async fn update_identity(&self, identity: &Identity) -> Result<()>;
-}
-
-/// Trait implementing read access to identiets
-#[async_trait]
-pub trait IdentitiesReader: Send + Sync + 'static {
-    /// Return a persisted identity
-    async fn retrieve_identity(&self, identifier: &Identifier) -> Result<Option<ChangeHistory>>;
-
-    /// Return a persisted identity that is expected to be present and return and Error if this is not the case
-    async fn get_identity(&self, identifier: &Identifier) -> Result<ChangeHistory> {
-        match self.retrieve_identity(identifier).await? {
-            Some(change_history) => Ok(change_history),
-            None => Err(Error::new(
-                Origin::Core,
-                Kind::NotFound,
-                format!("identity not found for identifier {}", identifier),
-            )),
-        }
-    }
-}
-
-/// Implementation of `IdentityAttributes` trait based on an underlying `Storage`
-#[derive(Clone)]
-pub struct IdentitiesStorage {
-    storage: Arc<dyn Storage>,
-    vault: Arc<dyn IdentitiesVault>,
 }
 
 impl IdentitiesStorage {
