+++ implementations/rust/ockam/ockam_identity/src/v2/identities/storage/mod.rs
@@ -1,21 +1,7 @@
 mod attributes_entry;
-mod identities_repository;
-/// LMDB implementation of the Storage trait
-#[cfg(feature = "std")]
-pub mod lmdb_storage;
-/// Sqlite implementation of the Storage trait
-#[cfg(feature = "sqlite")]
-pub mod sqlite_storage;
-
-#[allow(clippy::module_inception)]
-mod storage;
+mod identities_repository_impl;
+mod identities_repository_trait;
 
 pub use attributes_entry::*;
-pub use identities_repository::*;
-
-#[cfg(feature = "std")]
-pub use lmdb_storage::*;
-pub use storage::*;
-
-#[cfg(feature = "sqlite")]
-pub use sqlite_storage::*;
+pub use identities_repository_impl::*;
+pub use identities_repository_trait::*;
