+++ implementations/rust/ockam/ockam_identity/src/v2/mod.rs
@@ -30,6 +30,9 @@ pub mod secure_channel;
 /// Service supporting the creation of secure channel listener and connection to a listener
 pub mod secure_channels;
 
+/// Storage functions
+pub mod storage;
+
 ///
 /// Exports
 ///
