+++ implementations/rust/ockam/ockam_identity/src/v2/purpose_keys/mod.rs
@@ -2,3 +2,6 @@
 mod purpose_keys;
 
 pub use purpose_keys::*;
+
+/// Purpose Keys storage functions
+pub mod storage;
