+++ implementations/rust/ockam/ockam_identity/src/v2/purpose_keys/storage/mod.rs
@@ -1,0 +1,3 @@
+// mod purpose_keys_repository;
+//
+// pub use purpose_keys_repository::*;
